Reorganize conversation classes: put everything conversatoin related under Conversation namespace
