+++ lib/langchain/conversation/ai_message.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Langchain
+  class Conversation
+    class AIMessage < Message
+      def type
+        "ai"
+      end
+    end
+  end
+end
