+++ lib/langchain/conversation/human_message.rb
@@ -1,0 +1,11 @@
+# frozen_string_literal: true
+
+module Langchain
+  class Conversation
+    class HumanMessage < Message
+      def type
+        "human"
+      end
+    end
+  end
+end
