--- lib/langchain/system_message.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Langchain
-  class SystemMessage < Message
-    def type
-      "system"
-    end
-  end
-end
